add kitchen counter graph of counties

+++ demogState.h
@@ -15,11 +15,11 @@
 class demogState {
   public:
    //FILL IN
-   demogState(string inS, double in65, double in18,
-       double in5, int totPop20, Ethnicity ethnicity,
-       double mIncome, double homeowners, double pph, double vet,
-       double highSchool, double bachelors, double foreign, double hUnits,
-       double females, int counties) :
+   demogState(string inS, int in65, int in18,
+       int in5, int totPop20, Ethnicity ethnicity,
+       int mIncome, int homeowners, int pph, int vet,
+       int highSchool, int bachelors, int foreign, int hUnits,
+       int females, int counties) :
            state(inS), popOver65(in65), popUnder18(in18),
            popUnder5(in5), totalPopulation2020(totPop20),
            ethnicity(ethnicity), medianIncome(mIncome), homeowners(homeowners),
@@ -33,58 +33,57 @@ class demogState {
    //NOTE 1: these values are the counts from demogData NOT percentages
 
    //Population Info
-   double getPopOver65() const { return popOver65; }
-   double getPopOver65P() const { return (popOver65 / totalPopulation2020) * 100; }
+   int getPopOver65() const { return popOver65; }
+   double getPopOver65P() const { return (popOver65 / double(totalPopulation2020)) * 100; }
 
-   double getPopUnder18() const { return popUnder18; }
-   double getPopUnder18P() const { return (popUnder18 / totalPopulation2020) * 100; }
+   int getPopUnder18() const { return popUnder18; }
+   double getPopUnder18P() const { return (popUnder18 / double(totalPopulation2020)) * 100; }
 
-   double getPopUnder5() const { return popUnder5; }
-   double getPopUnder5P() const { return (popUnder5 / totalPopulation2020) * 100; }
+   int getPopUnder5() const { return popUnder5; }
+   double getPopUnder5P() const { return (popUnder5 / double(totalPopulation2020)) * 100; }
 
    int getTotalPopulation2020() const { return totalPopulation2020; }
 
    //Education Info
-   double getHighSchoolDegree() const { return highSchoolDegree; }
-   double getHighSchoolDegreeP() const { return (highSchoolDegree / totalPopulation2020) * 100; }
+   int getHighSchoolDegree() const { return highSchoolDegree; }
+   double getHighSchoolDegreeP() const { return (highSchoolDegree / double(totalPopulation2020)) * 100; }
 
-   double getBachelorsDegree() const { return bachelorsDegree; }
-   double getBachelorsDegreeP() const { return (bachelorsDegree / totalPopulation2020) * 100; }
+   int getBachelorsDegree() const { return bachelorsDegree; }
+   double getBachelorsDegreeP() const { return (bachelorsDegree / double(totalPopulation2020)) * 100; }
 
    //Housing info
-   double getHomeowners() const { return homeowners; }
-   double getHomeownersP() const { return (homeowners / totalPopulation2020) * 100; }
-   double getPersonsPerHouse() const { return personsPerHouse; }
-   double getHousingUnits() const { return housingUnits; }
+   int getHomeowners() const { return homeowners; }
+   double getHomeownersP() const { return (homeowners / double(totalPopulation2020)) * 100; }
+   int getPersonsPerHouse() const { return personsPerHouse; }
+   int getHousingUnits() const { return housingUnits; }
 
    //Miscellaneous Info
-   double getForeignBorn() const { return foreignBorn; }
-   double getVeterans() const { return veterans; }
-   double getMedianIncome() const { return medianIncome; }
-   double getFemales() const { return females; }
-   double getFemalesP() const { return (females / totalPopulation2020) * 100; }
+   int getForeignBorn() const { return foreignBorn; }
+   int getVeterans() const { return veterans; }
+   int getMedianIncome() const { return medianIncome; }
+   int getFemales() const { return females; }
+   double getFemalesP() const { return (females / double(totalPopulation2020)) * 100; }
    int getCounties() const { return counties; }
 
   friend std::ostream& operator<<(std::ostream &out, const demogState &SD);
 
   private:
-	//DATA here
   const string state;
-  const double popOver65;
-  const double popUnder18;
-  const double popUnder5;
+  const int popOver65;
+  const int popUnder18;
+  const int popUnder5;
   const int totalPopulation2020;
 
   const Ethnicity ethnicity;
-  const double medianIncome;
-  const double homeowners;
-  const double personsPerHouse;
-  const double veterans;
-  const double highSchoolDegree;
-  const double bachelorsDegree;
-  const double foreignBorn;
-  const double housingUnits;
-  const double females;
+  const int medianIncome;
+  const int homeowners;
+  const int personsPerHouse;
+  const int veterans;
+  const int highSchoolDegree;
+  const int bachelorsDegree;
+  const int foreignBorn;
+  const int housingUnits;
+  const int females;
   const int counties;
 
 };
